Add log to check tiering

+++ src/execution/arm64/simulator-arm64.h
@@ -2564,7 +2564,7 @@ class Simulator : public DecoderVisitor, public SimulatorBase {
 
   int log_parameters_;
   // Instruction counter only valid if v8_flags.stop_sim_at isn't 0.
-  int icount_for_stop_sim_at_;
+  int64_t icount_for_stop_sim_at_;
   Isolate* isolate_;
 };
